requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.1.1
cryptography==43.0.3
cycler==0.12.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
dotenv==0.9.9
fonttools==4.54.1
google_search_results==2.4.2
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lime==0.2.0.1
matplotlib==3.7.5
networkx==3.1
newsapi-python==0.2.7
numpy==1.24.4
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyWavelets==1.4.1
requests==2.32.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tifffile==2023.7.10
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
